Squashed commit of the following:
commit 9284e96451dfed34229f6554a1e46e0e6547bd7b
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 15:03:14 2025 -0500

    ships do not carry over between resets

commit 58a6f32547dd5a32d2a490ea1122c84417ba3c9c
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 14:22:22 2025 -0500

    fixed double game id on reset

commit 4c5b148c5ff5463d7e02194b1de2c351b47b81ec
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 14:12:41 2025 -0500

    createGame takes ship data

commit 0ad64b855a22df55541bba4969bfa912c556d1d6
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 13:52:10 2025 -0500

    cognito lambda deprecated

commit 69e78ac592308bd4ed3682e46f8e8dcb54eb9e6e
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 13:51:59 2025 -0500

    new websocket

commit 8c52ac0463dc27ddf1cb068edd01d0ce42abf1d8
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 13:51:51 2025 -0500

    update

commit 80d06df0931047334dc61cc608ba83e8c4929c96
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 13:51:45 2025 -0500

    sender.js fixed

commit 3d609a709a28928297a5123258510a6ec7cf9833
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 13:51:35 2025 -0500

    delete game fixed

commit 459c17d5cfe2dbfef320c679941b915eba03ef91
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 13:51:22 2025 -0500

    update

commit 64636cf3f4cdad71e88049f521f0112c6d38c826
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 13:51:06 2025 -0500

    docs

commit c9947626800c1c0ccfe20e47e582f16647d7d1c9
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 13:45:32 2025 -0500

    create game lambda fixed

commit 6f5e00b715c261c69206935124df2158a037e852
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 10:45:09 2025 -0500

    cognito lambda removed

commit ff79cd336d8fd12e8bd913f2024d41ea300a68f3
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 10:30:02 2025 -0500

    removed deprecated ship.js file

commit c3918465cbd47c0a1e500e4a4766885bd3940243
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 10:25:12 2025 -0500

    tutorial file cleaned

commit a812bd5b77bb84e3b762f38063a8c509308b16e8
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 10:21:32 2025 -0500

    tutorial fixed

commit 87049bb90c81bb1fa3163fd42d695a979fc7416e
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 10:07:20 2025 -0500

    docs

commit bcaa3f55bb154deb93bc5e69ccad4c5146c525a5
Author: maxlink <max.durham.link@gmail.com>
Date:   Tue Apr 15 09:47:39 2025 -0500

    root documenting

commit d4ce5cd053d38f0fdd73299313d7c7f9d0663bbe
Author: justinv04 <njkvincent@gmail.com>
Date:   Mon Apr 14 23:01:19 2025 -0500

    refactored unselected lambdas

commit b17cd05f2a0e2d34bedd880d78336ddf50db30c1
Merge: bf374b6 32ed470
Author: justinv04 <njkvincent@gmail.com>
Date:   Mon Apr 14 21:51:39 2025 -0500

    Merge branch 'documenting' of https://github.com/MaxDLink/CS5-7319FinalProjectGroup3-Link-Masters-Vincent into documenting

commit bf374b6da8f13279d91e65f0bf729b0e99bb36e2
Author: justinv04 <njkvincent@gmail.com>
Date:   Mon Apr 14 21:51:23 2025 -0500

    Refactored more of src I left out on accident

commit 32ed470a52a4ec56533633e2bce5565a39ff9cc4
Author: maxlink <max.durham.link@gmail.com>
Date:   Mon Apr 14 20:41:39 2025 -0500

    ship placement noise restored

commit 31539e2ea473d14e7e0f92f7861265e33ba7a1c8
Author: maxlink <max.durham.link@gmail.com>
Date:   Mon Apr 14 20:37:00 2025 -0500

    ship placement noise restored

commit 9de0a26e26e66f27b9dc2bf420ebd865ea27dbda
Author: maxlink <max.durham.link@gmail.com>
Date:   Mon Apr 14 18:27:59 2025 -0500

    winner popup import added to selected

commit c25e064e437c64a0434ee55552a9be87fc558eb8
Author: maxlink <max.durham.link@gmail.com>
Date:   Mon Apr 14 18:18:33 2025 -0500

    fixed s3 origin error with correct import & removed unused cognito import

commit e0ff1e824a3978e3b8435df6a0b04482b9cef352
Author: justinv04 <njkvincent@gmail.com>
Date:   Mon Apr 14 13:12:51 2025 -0500

    Refactored chatbox, enemy-ais, game-board, and gameboard-mixins

commit dd0e2619c8f3562f2be8725abac515b14c3f09bd
Author: justinv04 <njkvincent@gmail.com>
Date:   Mon Apr 14 12:34:39 2025 -0500

    refactored lib/App

commit 38f8e725481271fa81bebfc60a353ed31298f556
Author: justinv04 <njkvincent@gmail.com>
Date:   Sun Apr 13 23:21:40 2025 -0500

    more stuff

commit b716325644190d96823653df4dcff83858975527
Author: justinv04 <njkvincent@gmail.com>
Date:   Sun Apr 13 23:03:22 2025 -0500

    More comments and refactoring

commit 995d669d20fa5909a94339ab82981fa23e588b3b
Author: justinv04 <njkvincent@gmail.com>
Date:   Sun Apr 13 22:55:37 2025 -0500

    refactored game-processor and get-game

commit a76f6f6da80d024523df38444ce009200b74ec54
Author: justinv04 <njkvincent@gmail.com>
Date:   Sun Apr 13 22:24:31 2025 -0500

    Fixed comments for delete-game and event-publisher

commit 0506c2dfaade18b3bcc051850a8352e0e1a3da7a
Author: justinv04 <njkvincent@gmail.com>
Date:   Sun Apr 13 21:52:50 2025 -0500

    Fixed Comments for cognito/lambda

commit efc71ab06ddb171c2851be9398133794b497c89d
Author: justinv04 <you@example.com>
Date:   Sun Apr 13 16:36:57 2025 -0500

    Refactored bin/App and lambda/create-game

commit b4081632190627446f06edb4bbd100effb23dd38
Author: MaxDLink <max.durham.link@gmail.com>
Date:   Sun Apr 13 13:39:03 2025 -0500

    readme's reconfigured

+++ selected/lambda/delete-game/index.js
@@ -6,8 +6,9 @@ exports.handler = async (event) => {
   try {
     console.log('Event received:', JSON.stringify(event, null, 2));
     
-    // Extract details from the event
-    const detail = event.detail;
+    // Parse the detail from the EventBridge event
+    const detail = typeof event.detail === 'string' ? JSON.parse(event.detail) : event.detail;
+    
     const connectionId = detail.connectionId;
     const gameId = detail.gameId;
     
@@ -17,8 +18,15 @@ exports.handler = async (event) => {
         body: JSON.stringify({ message: 'Game ID is required' })
       };
     }
+
+    if (!connectionId) {
+      return {
+        statusCode: 400,
+        body: JSON.stringify({ message: 'Connection ID is required' })
+      };
+    }
     
-    // Delete the game from DynamoDB
+    // Database Interaction
     await dynamoDB.delete({
       TableName: process.env.DYNAMODB_TABLE,
       Key: {
@@ -29,7 +37,7 @@ exports.handler = async (event) => {
     
     console.log(`Game ${gameId} deleted successfully`);
     
-    // Publish an event to notify about the game deletion
+    // Publish to EventBridge
     await eventBridge.putEvents({
       Entries: [{
         Source: 'game.service',
@@ -50,9 +58,9 @@ exports.handler = async (event) => {
         gameId: gameId
       })
     };
-  } catch (error) {
+  }
+  catch (error) {
     console.error('Error deleting game:', error);
-    
     return {
       statusCode: 500,
       body: JSON.stringify({ 
@@ -62,10 +70,3 @@ exports.handler = async (event) => {
     };
   }
 };
-
-const corsHeaders = () => ({
-  'Access-Control-Allow-Origin': '*',
-  'Access-Control-Allow-Methods': 'DELETE, OPTIONS',
-  'Access-Control-Allow-Headers': 'Content-Type',
-  'Content-Type': 'application/json'
-});
